sensor_app/requirements.txt
grpcio
grpcio-tools
grpcio-health-checking
opentelemetry-sdk
opentelemetry-instrumentation-grpc
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp
